requirements.txt
# PikoPy
lxml

# Flask
Flask==1.1.1
Werkzeug==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Flask-Caching>=1.0.0

# Database
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15

# Cronjob
python-crontab==2.4.1

# Sunset
pytz
astral